fix: requirements削除を禁止

+++ UVR.py
@@ -2274,6 +2274,9 @@ class MainWindow(TkinterDnD.Tk if is_dnd_compatible else tk.Tk):
             for dir in DIRECTORIES:
                 for temp_file in os.listdir(dir):
                     if temp_file.endswith(EXTENSIONS):
+                        # Skip requirements files
+                        if temp_file.startswith('requirements'):
+                            continue
                         if os.path.isfile(os.path.join(dir, temp_file)):
                             os.remove(os.path.join(dir, temp_file))
         except Exception as e:
